+++ shmc/memory-pool.h
@@ -51,4 +51,4 @@ void *RegisterMemory(uint16_t id, uint32_t size);
 void *AcquireMemory(uint16_t id);
 void ReleaseMemory(uint16_t id);
 void ReleaseMemoryAndRollback (uint16_t id);
-uint8_t GetMemoryVersion(uint8_t id);
+uint8_t GetMemoryVersion(uint16_t id);
